+++ FO-DICOM.Core/DicomUIDGenerated.cs
@@ -1,5 +1,11 @@
-﻿// Copyright (c) 2012-2025 fo-dicom contributors.
+﻿
+// Copyright (c) 2012-2025 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
 
 namespace FellowOakDicom
 {